+++ src/thread/Thread.php
@@ -38,16 +38,6 @@ use pocketmine\scheduler\AsyncTask;
 abstract class Thread extends NativeThread{
 	use CommonThreadPartsTrait;
 
-	public function start(int $options = NativeThread::INHERIT_NONE) : bool{
-		//this is intentionally not traitified
-		ThreadManager::getInstance()->add($this);
-
-		if($this->getClassLoaders() === null){
-			$this->setClassLoaders();
-		}
-		return parent::start($options);
-	}
-
 	/**
 	 * Stops the thread using the best way possible. Try to stop it yourself before calling this.
 	 */
