+++ src/thread/Worker.php
@@ -23,7 +23,6 @@ declare(strict_types=1);
 
 namespace pocketmine\thread;
 
-use pmmp\thread\Thread as NativeThread;
 use pmmp\thread\Worker as NativeWorker;
 use pocketmine\scheduler\AsyncTask;
 
@@ -39,16 +38,6 @@ use pocketmine\scheduler\AsyncTask;
  */
 abstract class Worker extends NativeWorker{
 	use CommonThreadPartsTrait;
-
-	public function start(int $options = NativeThread::INHERIT_NONE) : bool{
-		//this is intentionally not traitified
-		ThreadManager::getInstance()->add($this);
-
-		if($this->getClassLoaders() === null){
-			$this->setClassLoaders();
-		}
-		return parent::start($options);
-	}
 
 	/**
 	 * Stops the thread using the best way possible. Try to stop it yourself before calling this.
